refactor: reorganize app into feature-based source layout
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
