+++ src/pages/NewBuildPage.tsx
@@ -1,6 +1,6 @@
-import type { BuildDraft, BuildSpec } from "@/lib/types";
+import type { BuildDraft, BuildSpec } from "@/shared/lib/types";
 
-export default function NewBuildPage() {
+export function NewBuildPage() {
   const initialSpec: BuildSpec = {
     datasetId: "",
     title: "",
@@ -18,7 +18,7 @@ export default function NewBuildPage() {
 
   return (
     <main className="mx-auto flex w-full max-w-4xl flex-1 flex-col gap-4 px-6 py-12">
-      <h1 className="text-3xl font-semibold tracking-tight">New Build</h1>
+      <h2 className="text-3xl font-semibold tracking-tight">New Build</h2>
       <p className="text-zinc-600 dark:text-zinc-300">
         Placeholder wizard for drafting, validating, and executing a new build
         specification.
